Split Smart Search preload progress
Move active preload progress persistence and window events into a small helper so indexer.ts keeps its attention on crawling OPAL content.
